Move DirectoryPathProvider to Data file
